Mark `NSUbiquitousKeyValueStoreStorage` internal
`Persisted` and `Persister` extensions have been removed.

Tests have been update to better split testing across types.
